追加

+++ src/adapterPattern/GlobalCalculate.java
@@ -1,5 +1,5 @@
 package adapterPattern;
 
 public interface GlobalCalculate {
-	public abstract int globalCalculate();
+	public abstract double globalCalculate();
 }
